titan/requirements.txt
torch>=2.0
transformers>=4.30
peft>=0.6
tiktoken>=0.5
pyyaml>=6.0
